Added the Weapon to both of Player's Hands

+++ Assets/Scripts/Character/Player/Player UI/PlayerUIHudManager.cs
@@ -15,7 +15,7 @@ public class PlayerUIHudManager : MonoBehaviour
         staminaBar.gameObject.SetActive(true);
     }
 
-    public void SetNewHealthValue(float oldValue, float newValue)
+    public void SetNewHealthValue(int oldValue, int newValue)
     {
         Debug.Log("SetNewHealthValue = Mathf.RoundToInt(newValue): " + Mathf.RoundToInt(newValue));
         healthBar.SetStat(Mathf.RoundToInt(newValue));
